rename: `@ExtendWith` 전용 클래스 네이밍 수정
