+++ src/test/java/com/kgu/studywithme/common/AcceptanceTest.java
@@ -7,8 +7,8 @@ import com.kgu.studywithme.auth.infrastructure.oauth.kakao.KakaoOAuthUriGenerato
 import com.kgu.studywithme.auth.infrastructure.oauth.naver.NaverOAuthConnector;
 import com.kgu.studywithme.auth.infrastructure.oauth.naver.NaverOAuthUriGenerator;
 import com.kgu.studywithme.common.config.ExternalApiConfiguration;
-import com.kgu.studywithme.common.config.MySqlTestContainersConfiguration;
-import com.kgu.studywithme.common.config.RedisTestContainersConfiguration;
+import com.kgu.studywithme.common.config.MySqlTestContainersExtension;
+import com.kgu.studywithme.common.config.RedisTestContainersExtension;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
@@ -21,8 +21,8 @@ import org.springframework.context.annotation.Import;
 @Tag("Acceptance")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ExtendWith({
-        MySqlTestContainersConfiguration.class,
-        RedisTestContainersConfiguration.class
+        MySqlTestContainersExtension.class,
+        RedisTestContainersExtension.class
 })
 @Import(ExternalApiConfiguration.class)
 public abstract class AcceptanceTest {
