+++ src/test/java/com/kgu/studywithme/common/config/DatabaseCleanerEachCallbackExtension.java
@@ -1,10 +1,11 @@
-package com.kgu.studywithme.common.utils;
+package com.kgu.studywithme.common.config;
 
+import com.kgu.studywithme.common.utils.DatabaseCleaner;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-public class DatabaseCleanerEachCallback implements AfterEachCallback {
+public class DatabaseCleanerEachCallbackExtension implements AfterEachCallback {
     @Override
     public void afterEach(final ExtensionContext context) {
         final DatabaseCleaner databaseCleaner = SpringExtension.getApplicationContext(context).getBean(DatabaseCleaner.class);
