+++ src/test/java/com/kgu/studywithme/common/config/RedisTestContainersExtension.java
@@ -5,7 +5,7 @@ import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers
-public class RedisTestContainersConfiguration implements Extension {
+public class RedisTestContainersExtension implements Extension {
     private static final String REDIS_IMAGE = "redis:latest";
     private static final int REDIS_PORT = 6379;
     private static final GenericContainer<?> CONTAINER;
